Fix ECR Controller Bug - Ensuring Proper Resource Reconciliation on Repository Creation
**TLDR:**
- Addressed a bug in the ECR controller where Repository Custom Resources (CRs) failed
  to create Policy and Lifecycle Policy entries when containing all fields.
- This issue led to proper reconciliation only when a resource was requeued. It occurred
  primarily due to late initialization, detecting API default values for non-specified
  fields, causing the controller to requeue resources after they've been created.
- Resolved by adding the necessary hooks to the create path.

**Description:**
This fix addresses a bug in the ECR controller. When a user submitted a Repository
Custom Resource (CR) containing all fields, the controller failed to create Policy
and Lifecycle Policy entries. However, if one of the fields was missing (excluding
`Policy` or `LifecyclePolicy`), the reconciliation process succeeded.

The bug fix was straightforward and intuitive, as it became evident that the controller
lacked the required hook for the create path. However, this raised questions about why
`policy` and `LifecyclePolicy` were missing in specific cases. Our end-to-end (e2e) tests
created CRs with isolated fields, such as `Policy` and `LifecyclePolicy`, and verified
that these policies were indeed applied to the ECR Repository. The tests failed when the
policies were missing.

It became evident that the reason these specific cases were successfully reconciled was
due to the correct logic in the update path. This logic ensured that if a `Policy` or
`LifecyclePolicy` was missing, an API call was made to update these fields.

The mystery revolved around why these edge cases were reconciled even when we didn't
explicitly request a second reconciliation or requeue of the resource.

- Our debugging journey began with a review of the logs and a comparison between the logs
  of a resource created with all fields and one with only the policy. Surprisingly, the
  logs showed a complete reconciliation step, and it wasn't initiated by the controller
  itself (no requeue, no drift remediation, no condition). This was puzzling.

- To gain deeper insights, we connected the controller to a local clone of the runtime and
  added print statements. However, this approach made the logs messier and harder to
  interpret. Consequently, we turned to debugging with a debugger.

- Our debugging session took us through the AWS Controllers for Kubernetes (ACK) runtime,
  but the issue didn't originate there. Instead, it stemmed from 'k8s-sigs/controller-runtime.'
  We cloned it and extended our debugging to the controller-runtime. There, we discovered
  that 'c.Do.Reconcile(ctx, req)' was causing two reconciliations.

- Further investigation revealed that 'for c.processNextWorkItem(ctx) {}' was called,
  indicating the presence of an extra item in the queue. This item had to be an update
  because the first item in the queue was a create.

- To delve deeper, we created two resources—one with all fields and the other with only
  the name and policy. We then ran a 'diff' on 'k get repository <name> -o yaml,' which
  revealed a crucial clue: 'metadata.generation' was 1 for the all-fields resource and 2
  for the policy+name-only resource, confirming the update theory.

- The controller initiated the resource update because the ECR API returned all fields,
  even those not explicitly supplied. This behavior prompted the ACK runtime to execute
  a patch into the resource spec, incrementing the generation and triggering an update.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>